vastly improve prebuild script
