+++ packages/manager/package.json
@@ -92,7 +92,7 @@
     "start": "vite",
     "start:ci": "yarn serve ./build -p 3000 -s --cors",
     "lint": "yarn run eslint . --ext .js,.ts,.tsx --quiet",
-    "build": "node scripts/build.js && vite build",
+    "build": "node scripts/prebuild.mjs && vite build",
     "precommit": "lint-staged && yarn typecheck",
     "test": "jest --color",
     "test:debug": "node --inspect-brk scripts/test.js --runInBand",
@@ -171,7 +171,6 @@
     "axe-core": "^3.5.1",
     "browserslist": "^4.16.5",
     "chai-string": "^1.5.0",
-    "chalk": "^2.4.2",
     "core-js": "^2.6.5",
     "cypress": "^12.2.0",
     "cypress-axe": "^1.0.0",
